added: ExceptionHandler controller

+++ src/main/java/com/vbaggio/ecom/services/ProductService.java
@@ -1,15 +1,19 @@
 package com.vbaggio.ecom.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.vbaggio.ecom.dto.ProductDTO;
 import com.vbaggio.ecom.entitites.Product;
 import com.vbaggio.ecom.mappers.ProductMapper;
 import com.vbaggio.ecom.repositories.ProductRepository;
+import com.vbaggio.ecom.services.exceptions.DatabaseException;
+import com.vbaggio.ecom.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ProductService {
@@ -17,9 +21,14 @@ public class ProductService {
 	@Autowired
 	private ProductRepository repository;
 
+	private Product entityFromId(Long id) {
+		return repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Product not found with id " + id));
+	}
+
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product result = repository.findById(id).get();
+		Product result = entityFromId(id);
 		return ProductMapper.INSTANCE.toDto(result);
 	}
 
@@ -38,15 +47,23 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.findById(id).get();
+		Product entity = entityFromId(id);
 		ProductMapper.INSTANCE.updateEntityFromDTO(dto, entity);
 		entity = repository.save(entity);
 		return ProductMapper.INSTANCE.toDto(entity);
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Product not found with id " + id);
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException();
+		}
+		
 	}
 
 }
